requirements/requirements.txt
django==2.1.4
djangorestframework==3.9.0
pillow==5.3.0
